Using new payment api
